Move entry files to root directory
